+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/DependencyManager.groovy
@@ -1,24 +1,30 @@
 package life.qbic.portal.sampletracking
 
-import com.vaadin.icons.VaadinIcons
-import com.vaadin.ui.Button
 import com.vaadin.ui.VerticalLayout
 import life.qbic.business.project.load.LoadProjects
 import life.qbic.business.project.load.LoadProjectsDataSource
 import life.qbic.business.project.load.LoadProjectsInput
 import life.qbic.business.project.load.LoadProjectsOutput
+import life.qbic.business.samples.count.CountSamples
+import life.qbic.business.samples.count.CountSamplesDataSource
+import life.qbic.business.samples.count.CountSamplesOutput
 import life.qbic.datamodel.dtos.portal.PortalUser
 import life.qbic.datamodel.dtos.projectmanagement.Project
 import life.qbic.portal.sampletracking.communication.notification.MessageBroker
 import life.qbic.portal.sampletracking.communication.notification.NotificationService
 import life.qbic.portal.sampletracking.components.NotificationHandler
+import life.qbic.portal.sampletracking.components.projectoverview.CountSamplesPresenter
 import life.qbic.portal.sampletracking.components.projectoverview.LoadProjectsPresenter
 import life.qbic.portal.sampletracking.components.projectoverview.ProjectOverviewView
 import life.qbic.portal.sampletracking.components.projectoverview.ProjectOverviewViewModel
 import life.qbic.portal.sampletracking.datasources.Credentials
 import life.qbic.portal.sampletracking.datasources.OpenBisConnector
+import life.qbic.portal.sampletracking.datasources.database.DatabaseSession
+import life.qbic.portal.sampletracking.datasources.samples.SamplesDbConnector
 import life.qbic.portal.sampletracking.resource.ResourceService
 import life.qbic.portal.sampletracking.resource.project.ProjectResourceService
+import life.qbic.portal.sampletracking.resource.status.StatusCount
+import life.qbic.portal.sampletracking.resource.status.StatusCountResourceService
 import life.qbic.portal.utils.ConfigurationManager
 import life.qbic.portal.utils.ConfigurationManagerFactory
 
@@ -39,7 +45,10 @@ class DependencyManager {
     private final NotificationHandler notificationHandler
 
     private LoadProjectsDataSource loadProjectsDataSource
+    private CountSamplesDataSource countSamplesDataSource
+
     private ResourceService<Project> projectResourceService
+    private ResourceService<StatusCount> statusCountService
     private NotificationService notificationService
 
     DependencyManager(PortalUser user) {
@@ -52,6 +61,7 @@ class DependencyManager {
 
         populateProjectService()
         portletView = setupPortletView()
+        populateStatusCountService()
     }
 
 
@@ -62,10 +72,21 @@ class DependencyManager {
 
     private void setupServices() {
         projectResourceService = new ProjectResourceService()
+        statusCountService = new StatusCountResourceService()
         notificationService = new MessageBroker()
     }
 
     private void setupDatabaseConnections() {
+        String user = Objects.requireNonNull(configurationManager.getMysqlUser(), "Mysql user missing.")
+        String password = Objects.requireNonNull(configurationManager.getMysqlPass(), "Mysql password missing.")
+        String host = Objects.requireNonNull(configurationManager.getMysqlHost(), "Mysql host missing.")
+        String port = Objects.requireNonNull(configurationManager.getMysqlPort(), "Mysql port missing.")
+        String sqlDatabase = Objects.requireNonNull(configurationManager.getMysqlDB(), "Mysql database name missing.")
+
+        DatabaseSession.init(user, password, host, port, sqlDatabase)
+        SamplesDbConnector samplesDbConnector = new SamplesDbConnector(DatabaseSession.getInstance())
+        countSamplesDataSource = samplesDbConnector
+
         Credentials openBisCredentials = new Credentials(
                 user: configurationManager.getDataSourceUser(),
                 password: configurationManager.getDataSourcePassword()
@@ -91,12 +112,12 @@ class DependencyManager {
      * Creates a new ProjectOverviewView using
      * <ul>
      *     <li>{@link #projectResourceService}</li>
-     *     <li>{@link #loadProjectsDataSource}</li>
+     *     <li>{@link #statusCountService}</li>
      * </ul>
      * @return a new ProjectOverviewView
      */
     private ProjectOverviewView createProjectOverviewView() {
-        ProjectOverviewViewModel viewModel = new ProjectOverviewViewModel(projectResourceService)
+        ProjectOverviewViewModel viewModel = new ProjectOverviewViewModel(projectResourceService, statusCountService)
         ProjectOverviewView view =  new ProjectOverviewView(viewModel)
         return view
     }
@@ -112,6 +133,18 @@ class DependencyManager {
     }
 
     /**
+     * Triggers the project status count loading initially to have data in the service
+     */
+    private void populateStatusCountService() {
+        CountSamplesOutput output = new CountSamplesPresenter(notificationService, statusCountService)
+        CountSamples countSamples = new CountSamples(countSamplesDataSource, output)
+        List<String> projectCodes = projectResourceService.iterator().collect {
+            return it.projectId.projectCode.toString()
+        }
+        projectCodes.each {countSamples.countReceivedSamples(it)}
+    }
+
+        /**
      * Returns the global notification center
      * @return a notification center that handles app notifications
      */