+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectOverviewViewModel.groovy
@@ -3,6 +3,7 @@ package life.qbic.portal.sampletracking.components.projectoverview
 import life.qbic.datamodel.dtos.projectmanagement.Project
 import life.qbic.portal.sampletracking.communication.Topic
 import life.qbic.portal.sampletracking.resource.ResourceService
+import life.qbic.portal.sampletracking.resource.status.StatusCount
 
 /**
  * <h1>ViewModel for the {@link ProjectOverviewView}</h1>
@@ -14,22 +15,50 @@ import life.qbic.portal.sampletracking.resource.ResourceService
 */
 class ProjectOverviewViewModel {
 
-    ObservableList projects = new ObservableList(new ArrayList<Project>())
+    ObservableList projectOverviews = new ObservableList(new ArrayList<ProjectSummary>())
     private final ResourceService<Project> projectResourceService
+    private final ResourceService<StatusCount> statusCountService
 
-    ProjectOverviewViewModel(ResourceService<Project> projectResourceService){
+    ProjectOverviewViewModel(ResourceService<Project> projectResourceService, ResourceService<StatusCount> statusCountService){
         this.projectResourceService = projectResourceService
+        this.statusCountService = statusCountService
         fetchProjectData()
         subscribeToResources()
     }
 
     private void fetchProjectData() {
-        projects.clear()
-        projects.addAll(projectResourceService.iterator())
+        projectOverviews.clear()
+        projectResourceService.iterator().each {Project project ->
+            addProject(project)
+        }
+        statusCountService.iterator().each { StatusCount statusCount ->
+            updateSamplesReceived(statusCount.projectCode, statusCount.count)
+        }
     }
 
     private void subscribeToResources() {
-        this.projectResourceService.subscribe({ projects.add(it) }, Topic.PROJECT_ADDED)
-        this.projectResourceService.subscribe({ projects.remove(it) }, Topic.PROJECT_REMOVED)
+        this.projectResourceService.subscribe({ addProject(it) }, Topic.PROJECT_ADDED)
+        this.projectResourceService.subscribe({ removeProject(it) }, Topic.PROJECT_REMOVED)
+
+        this.statusCountService.subscribe({updateSamplesReceived(it.projectCode, it.count)}, Topic.SAMPLE_RECEIVED_COUNT_UPDATE)
+    }
+
+    private void addProject(Project project) {
+        ProjectSummary.Builder builder = new ProjectSummary.Builder(project)
+        projectOverviews.add(builder.build())
+    }
+
+    private void removeProject(Project project) {
+        ProjectSummary projectOverview = projectOverviews.collect {it as ProjectSummary}.find { it ->
+            (it as ProjectSummary).code == project.projectId.projectCode.toString()
+        }
+        projectOverviews.remove(projectOverview)
+    }
+
+    private void updateSamplesReceived(String projectCode, int sampleCount) {
+        ProjectSummary projectOverview = projectOverviews.collect {it as ProjectSummary}.find { it ->
+            (it as ProjectSummary).code == projectCode
+        }
+        projectOverview.samplesReceived = sampleCount
     }
 }